requirements.txt
asgiref==3.8.1
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
credentials==1.1
Django==5.1.2
django-admin==2.0.2
django-cors-headers==4.6.0
django-excel-response2==3.0.6
django-firebase-orm==0.6.1
django-rest-framework==0.1.0
django-six==1.0.5
djangorestframework==3.15.2
drf-yasg==1.21.8
excel-base==1.0.4
firebase-admin==2.13.0
gcloud==0.18.3
google-api-core==2.22.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
httplib2==0.22.0
idna==3.10
inflection==0.5.1
isoweek==1.3.3
jws==0.1.3
msgpack==1.1.0
oauth2client==4.1.3
packaging==24.1
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pyparsing==3.2.0
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jwt==2.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
screen==1.0.1
six==1.16.0
sqlparse==0.5.1
TimeConvert==3.0.13
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.20
whitenoise==6.8.2
xlwt==1.3.0
